perf

+++ src/plugins/intel_gpu/tests/functional/remote_tensor_tests/helpers.hpp
@@ -193,8 +193,8 @@ struct OpenCL {
 };
 
 inline std::shared_ptr<ov::Model> make_copy_model(const ov::Shape& shape) {
-    auto param = std::make_shared<ov::op::v0::Parameter>(ov::element::f32, shape);
-    auto zero = ov::op::v0::Constant::create(ov::element::f32, ov::Shape{1}, {0.0f});
+    auto param = std::make_shared<ov::op::v0::Parameter>(ov::element::f16, shape);
+    auto zero = ov::op::v0::Constant::create(ov::element::f16, ov::Shape{1}, {0.0f});
     auto add = std::make_shared<ov::op::v1::Add>(param, zero);
     auto result = std::make_shared<ov::op::v0::Result>(add);
     return std::make_shared<ov::Model>(ov::ResultVector{result}, ov::ParameterVector{param});
